refactoring: Adding response entity to routes

+++ src/main/java/dev/java10x/cadastroDeNinjas/Ninjas/NinjaController.java
@@ -1,6 +1,9 @@
 package dev.java10x.cadastroDeNinjas.Ninjas;
 
+import org.apache.coyote.Response;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -23,34 +26,57 @@ public class NinjaController {
 
     //adicionar ninja (CREATE)
     @PostMapping("/criar")
-    public NinjaDTO criarNinja(@RequestBody NinjaDTO ninjaDTO){
-        return ninjaService.criarNinja(ninjaDTO);
+    public ResponseEntity<String> criarNinja(@RequestBody NinjaDTO ninjaDTO){
+        NinjaDTO novoNinja = ninjaService.criarNinja(ninjaDTO);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Ninja criado com sucesso: " + novoNinja.getNome()
+                + " | ID: " + novoNinja.getId());
+
     }
 
     //mostrar todos os ninjas (READ)
     @GetMapping("/listar")
-    public List<NinjaDTO> listarNinjas(){
-        return ninjaService.listarNinjas();
+    public ResponseEntity<List<NinjaDTO>> listarNinjas(){
+        List<NinjaDTO> ninjas = ninjaService.listarNinjas();
+        return ResponseEntity.ok(ninjas);
     }
 
     // procurar ninja por id (READ)
     @GetMapping("/listar/{id}")
-    public NinjaDTO listarNinjasPorId(@PathVariable Long id){
-        return ninjaService.listarNinjasPorId(id);
+    public ResponseEntity<?> listarNinjasPorId(@PathVariable Long id){
+        NinjaDTO ninja = ninjaService.listarNinjasPorId(id);
+        if(ninja != null){
+            ninjaService.listarNinjasPorId(id);
+            return ResponseEntity.ok(ninja);
+        }
+        else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Ninja ID: " + id +  " não encontrado");
+        }
     }
 
     //alterar dado do ninja (UPDATE)
     @PutMapping("/alterar/{id}")
-    public String alterarNinjaPorId(@PathVariable Long id, @RequestBody NinjaDTO ninjaDTO){
-        ninjaService.atualizarNinja(id, ninjaDTO);
-        return "Ninja ID: " + id +" alterado";
+    public ResponseEntity<String> alterarNinjaPorId(@PathVariable Long id, @RequestBody NinjaDTO ninjaDTO){
+        if(ninjaService.listarNinjasPorId(id) != null){
+            ninjaService.atualizarNinja(id, ninjaDTO);
+            return ResponseEntity.ok().body("Ninja ID: " + id + " alterado");
+        }
+        else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Ninja ID: " + id + " não encontrado");
+        }
+
     }
 
     //deletar ninja (DELETE
    @DeleteMapping("/deletar/{id}")
-    public String deleteNinjaPorId(@PathVariable Long id) {
-        ninjaService.deleteNinjaPorId(id);
-        return "Ninja ID: "+ id + " deletado.";
+    public ResponseEntity<String> deleteNinjaPorId(@PathVariable Long id) {
+        if(ninjaService.listarNinjasPorId(id) != null){
+            ninjaService.deleteNinjaPorId(id);
+            return ResponseEntity.ok("Ninja #" + id + " deletado com sucesso");
+        }
+        else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Ninja ID: " + id + " nao encontrado");
+        }
     }
